添加添加股票操作记录的接口

+++ database/model/operation.go
@@ -8,7 +8,7 @@ type Operation struct {
 	AccountID  uint64
 	StockID    uint64
 	SharePrice float64   `gorm:"type:double"`
-	BuyNum     float64   `gorm:"type:double;default:0"`
-	SaleNum    float64   `gorm:"type:double;default:0"`
+	BuyNum     int       `gorm:"type:double;default:0"`
+	SaleNum    int       `gorm:"type:double;default:0"`
 	Time       time.Time `gorm:"type:datetime;default:'1000-01-01 00:00:00'"`
 }
